# Improve RuntimeHelpers.GetSubArray